parsefloat

+++ js/script.js
@@ -18,9 +18,9 @@ if (navigator.serviceWorker) {
  */
 function myButtonClicked() {
   // input
-  const length = parseInt(document.getElementById("length").value)
-  const width = parseInt(document.getElementById("width").value)
-  const height = parseInt(document.getElementById("height").value)
+  const length = parseFloat(document.getElementById("length").value)
+  const width = parseFloat(document.getElementById("width").value)
+  const height = parseFloat(document.getElementById("height").value)
 
   // process
   const volume = (length * width * height) / 3
